Create CITY TEA
The freshness of Iced Tea drinks is unmatched. It can be said that this tea drink is the favorite of Indonesian society. It can be seen from the habit, every morning you start the day with a glass of hot tea, don't forget to accompany your lunch with iced tea, and end the day with a cup of hot tea in the afternoon. However, tea is attached to my heart